+++ src/App.jsx
@@ -1,16 +1,30 @@
 import React from "react";
-import {View} from "react-native-windows";
-import ContentView from "./screens/contentView/ContentView";
-import LeftBar from "./screens/leftBar/LeftBar";
+import {NavigationContainer} from "@react-navigation/native";
+import MainStack from "./navigation/stacks/mainStack/MainStack";
+import NavBar from "./navigation/stacks/navBar/NavBar";
+import {enableScreens} from "react-native-screens";
+import {createStackNavigator} from "@react-navigation/stack";
+import {MAIN_STACK} from "./constants/routes";
+
+enableScreens();
+
+export const navigationRef = React.createRef();
+const Stack = createStackNavigator();
 
 const MainScreen = () => {
   return (
-    <View style={{height: "100%", backgroundColor: "#212223"}}>
-      <View style={{flexDirection: "row"}}>
-        <LeftBar />
-        <ContentView />
-      </View>
-    </View>
+    <>
+      <NavigationContainer ref={navigationRef}>
+        <NavBar navigator={navigationRef} />
+        <Stack.Navigator>
+          <Stack.Screen
+            name={MAIN_STACK}
+            component={MainStack}
+            options={{headerShown: false}}
+          />
+        </Stack.Navigator>
+      </NavigationContainer>
+    </>
   );
 };