+++ src/constants/routes.js
@@ -1,0 +1,5 @@
+export const MAIN_STACK = "MainStack";
+export const ITCHIO_STACK = "ItchioStack";
+export const ITCHIO_AUTH_ROUTE = "ItchioAuthRoute";
+export const PLAYDATE_STACK = "PlaydateStack";
+export const PLAYDATE_AUTH_ROUTE = "PlaydateAuthRoute";