+++ src/navigation/stacks/navBar/navBarRoutes.js
@@ -1,0 +1,17 @@
+import {
+  ITCHIO_AUTH_ROUTE,
+  PLAYDATE_AUTH_ROUTE,
+} from "../../../constants/routes";
+
+const navBarRoutes = [
+  {
+    route: ITCHIO_AUTH_ROUTE,
+    name: "Itchio",
+  },
+  {
+    route: PLAYDATE_AUTH_ROUTE,
+    name: "Playdate",
+  },
+];
+
+export default navBarRoutes;