+++ src/screens/itchioOauth/ItchioOauth.jsx
@@ -1,101 +1,31 @@
-import React, {useState} from "react";
+import React, {useCallback, useEffect, useState} from "react";
 import {TextInput} from "react-native";
-import {Alert, Button, Text, View} from "react-native-windows";
-import {asyncLogout, checkToken, signInAsync} from "../../helper/auth";
+import useItchioStore from "../../store/itchio";
+import {Button, Text, View} from "react-native-windows";
+import {signInAsync} from "../../helper/auth";
+import AsyncStorage from "@react-native-async-storage/async-storage";
+import BaseScreen from "../../components/baseScreen/BaseScreen";
 
 const ItchioOAuth = () => {
-  const [token, setToken] = useState(null);
-  const [awaiting, setAwait] = useState(false);
-  const [signedIn, setSignedIn] = useState(false);
-  const [info, setInfo] = useState(null);
+  const [tokenInput, setTokenInput] = useState();
+  const {token, logout, validateToken, awaitingToken, setAwait} =
+    useItchioStore();
 
-  const logout = () => {
-    asyncLogout();
-
-    setAwait(false);
-    setSignedIn(false);
-    setInfo(null);
-    setToken(null);
-  };
-
-  const submit = () => {
+  const startSignIn = () => {
     signInAsync();
     setAwait(true);
   };
 
-  const validateToken = async () => {
-    const result = await checkToken(token);
-    if (result) {
-      setSignedIn(true);
-      setInfo(result);
-    } else {
-      Alert.alert(
-        "Invalid Token",
-        "Your token isn't valid, please copy the providen token after authorizing the app",
-        [
-          {
-            text: "Cancel",
-            style: "cancel",
-          },
-          {
-            text: "Try Again",
-            onPress: () => {
-              setAwait(false);
-            },
-          },
-        ],
-      );
-    }
-  };
+  const asyncLogin = useCallback(async () => {
+    const accessToken = await AsyncStorage.getItem("userToken");
+    validateToken(accessToken);
+  }, []);
 
-  if (signedIn) {
-    return (
-      <View
-        style={{
-          flexDirection: "column",
-          backgroundColor: "#fa5c5c",
-          flexGrow: 0,
-          padding: 15,
-          margin: 7,
-          borderRadius: 14,
-        }}>
-        <Text>Success</Text>
-        <Text style={{marginVertical: 15}}>{info}</Text>
-        <Button title="Logout" onPress={() => logout()} />
-      </View>
-    );
-  }
+  useEffect(() => {
+    asyncLogin();
+  }, [asyncLogin]);
 
-  if (awaiting) {
-    return (
-      <View
-        style={{
-          flexDirection: "column",
-          backgroundColor: "#fa5c5c",
-          flexGrow: 0,
-          padding: 15,
-          margin: 7,
-          borderRadius: 14,
-        }}>
-        <Text>Enter Token</Text>
-        <TextInput
-          style={{borderColor: "black", borderRadius: 5, marginVertical: 5}}
-          onChangeText={token => setToken(token)}
-          onSubmitEditing={token => setToken(token)}
-          secureTextEntry={true}
-          value={token}
-          placeholder="token"
-        />
-        <Button
-          title="Validate Token"
-          onPress={() => validateToken()}
-          disabled={!token}
-        />
-      </View>
-    );
-  }
-
-  return (
+  const Success = () => (
     <View
       style={{
         flexDirection: "column",
@@ -105,11 +35,62 @@ const ItchioOAuth = () => {
         margin: 7,
         borderRadius: 14,
       }}>
-      <Text style={{color: "#222222", fontWeight: "600", marginBottom: 5}}>
-        Itchio OAuth
-      </Text>
-      <Button title="OAuth Login" onPress={submit} />
+      <Text>Success</Text>
+      <Button title="Logout" onPress={() => logout()} />
     </View>
+  );
+
+  const InputToken = () => (
+    <View
+      style={{
+        flexDirection: "column",
+        backgroundColor: "#fa5c5c",
+        flexGrow: 0,
+        padding: 15,
+        margin: 7,
+        borderRadius: 14,
+      }}>
+      <Text>Enter Token</Text>
+      <TextInput
+        style={{borderColor: "black", borderRadius: 5, marginVertical: 5}}
+        onChangeText={token => setTokenInput(token)}
+        onSubmitEditing={token => setTokenInput(token)}
+        secureTextEntry={true}
+        value={tokenInput}
+        placeholder="token"
+      />
+      <Button
+        title="Validate Token"
+        onPress={() => validateToken(tokenInput)}
+        disabled={!tokenInput}
+      />
+    </View>
+  );
+
+  return (
+    <BaseScreen centerContent>
+      <View style={{width: "50%"}}>
+        {token ? Success() : null}
+        {awaitingToken && !token ? InputToken() : null}
+        {!token && !awaitingToken ? (
+          <View
+            style={{
+              flexDirection: "column",
+              backgroundColor: "#fa5c5c",
+              flexGrow: 0,
+              padding: 15,
+              margin: 7,
+              borderRadius: 14,
+            }}>
+            <Text
+              style={{color: "#222222", fontWeight: "600", marginBottom: 5}}>
+              Itchio OAuth
+            </Text>
+            <Button title="OAuth Login" onPress={startSignIn} />
+          </View>
+        ) : null}
+      </View>
+    </BaseScreen>
   );
 };